Correct sort parameter in loop_and_display
As it turns out, the function always sorted by all columns
whether or not sort was TRUE.
It seems that it makes little sense to not sort
columns other than the inner-most (as then the
headings do not really fold into one another sensibly).

Therefore, the behaviour of sort is now to define
whether grouping on the inner-most variable is performed.